# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Simple wrapper around Azure DevOps API to provide consisentency of API response for Krateo Operator Generator (KOG)",
        "title": "Azure DevOps Plugin API for Krateo Operator Generator (KOG)",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Krateo Support",
            "url": "https://krateo.io",
            "email": "contact@krateoplatformops.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/{organization}/{project}/pipelines/": {
            "post": {
                "description": "Create a new Pipeline",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new Pipeline",
                "operationId": "post-pipeline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization name",
                        "name": "organization",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project ID or name",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API version (e.g., 7.2-preview.1)",
                        "name": "api-version",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Basic Auth header (Basic \u003cbase64-encoded-username:password\u003e)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Pipeline creation request body",
                        "name": "pipelineCreate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pipeline.CreatePipelineRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Pipeline details",
                        "schema": {
                            "$ref": "#/definitions/pipeline.CreatePipelineResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/api/{organization}/{project}/pipelines/{id}": {
            "get": {
                "description": "Get",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a pipeline",
                "operationId": "get-pipeline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization name",
                        "name": "organization",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project name or ID",
                        "name": "project",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pipeline ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API version (e.g., 7.2-preview.1)",
                        "name": "api-version",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Basic Auth header (Basic \u003cbase64-encoded-username:password\u003e)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pipeline details",
                        "schema": {
                            "$ref": "#/definitions/pipeline.GetPipelineResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "description": "Update a pipeline using build definitions endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a pipeline",
                "operationId": "put-pipeline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization name",
                        "name": "organization",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project name or ID",
                        "name": "project",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pipeline ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API version (e.g., 7.2-preview.1)",
                        "name": "api-version",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Basic Auth header (Basic \u003cbase64-encoded-username:password\u003e)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Pipeline update request body",
                        "name": "pipelineUpdate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pipeline.UpdatePipelineRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated pipeline details",
                        "schema": {
                            "$ref": "#/definitions/pipeline.UpdatePipelineResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "Delete a pipeline using build definitions endpoint",
                "summary": "Delete a pipeline",
                "operationId": "delete-pipeline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization name",
                        "name": "organization",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project name or ID",
                        "name": "project",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pipeline ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Basic Auth header (Basic \u003cbase64-encoded-username:password\u003e)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content - Pipeline deleted successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "pipeline.CreatePipelineConfigurationParametersComplete": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                },
                "repository": {
                    "description": "Required",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pipeline.BuildRepository"
                        }
                    ]
                },
                "type": {
                    "description": "Required - enum: unknown, yaml, designerJson, justInTime, designerHyphenJson",
                    "type": "string"
                }
            }
        },
        "pipeline.CreatePipelineRequest": {
            "type": "object",
            "properties": {
                "configuration": {
                    "description": "Required",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pipeline.CreatePipelineConfigurationParametersComplete"
                        }
                    ]
                },
                "folder": {
                    "type": "string"
                },
                "name": {
                    "description": "Required",
                    "type": "string"
                }
            }
        },
        "pipeline.CreatePipelineResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "$ref": "#/definitions/pipeline.ReferenceLinks"
                },
                "configuration": {
                    "$ref": "#/definitions/pipeline.PipelineConfiguration"
                },
                "folder": {
                    "description": "Embedded fields from PipelineBase",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "revision": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "pipeline.GetPipelineResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "$ref": "#/definitions/pipeline.ReferenceLinks"
                },
                "configuration": {
                    "$ref": "#/definitions/pipeline.PipelineConfiguration"
                },
                "folder": {
                    "description": "Embedded fields from PipelineBase",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "revision": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "pipeline.ReferenceLinks": {
            "type": "object",
            "properties": {
                "links": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "pipeline.UpdatePipelineRequest": {
            "type": "object",
            "properties": {
                "badgeEnabled": {
                    "type": "boolean"
                },
                "buildNumberFormat": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "Required",
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "process": {},
                "processParameters": {
                    "$ref": "#/definitions/pipeline.ProcessParameters"
                },
                "properties": {
                    "type": "object",
                    "additionalProperties": true
                },
                "quality": {
                    "type": "string"
                },
                "repository": {
                    "$ref": "#/definitions/pipeline.BuildRepository"
                },
                "revision": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                },
                "variableGroups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pipeline.VariableGroup"
                    }
                },
                "variables": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/pipeline.BuildDefinitionVariable"
                    }
                }
            }
        },
        "pipeline.UpdatePipelineResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "$ref": "#/definitions/pipeline.ReferenceLinks"
                },
                "badgeEnabled": {
                    "type": "boolean"
                },
                "buildNumberFormat": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "createdDate": {
                    "type": "string"
                },
                "dateCreated": {
                    "type": "string"
                },
                "draftOf": {
                    "$ref": "#/definitions/pipeline.BuildDefinitionReference"
                },
                "drafts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pipeline.BuildDefinitionReference"
                    }
                },
                "dropLocation": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "jobAuthorizationScope": {
                    "type": "string"
                },
                "jobCancelTimeoutInMinutes": {
                    "type": "integer"
                },
                "jobTimeoutInMinutes": {
                    "type": "integer"
                },
                "latestBuild": {
                    "$ref": "#/definitions/pipeline.Build"
                },
                "latestCompletedBuild": {
                    "$ref": "#/definitions/pipeline.Build"
                },
                "metrics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pipeline.BuildMetric"
                    }
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "process": {},
                "processParameters": {
                    "$ref": "#/definitions/pipeline.ProcessParameters"
                },
                "project": {
                    "$ref": "#/definitions/pipeline.ProjectReference"
                },
                "properties": {
                    "type": "object",
                    "additionalProperties": true
                },
                "quality": {
                    "type": "string"
                },
                "repository": {
                    "$ref": "#/definitions/pipeline.BuildRepository"
                },
                "revision": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                },
                "uri": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "variableGroups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pipeline.VariableGroup"
                    }
                },
                "variables": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/pipeline.BuildDefinitionVariable"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "basic"
        }
    },
...
}